backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python-headless
Pillow
python-multipart
